Fixed two issues related to validation of message integrity, nonce & fp.
Fixed the two pending issues related to validation of TURN request
messages.

1. In TURNS module when a request is received, it needs to be validated for
a number of parameters such as presence of username, realm etc. However,
the validation of Nonce, Message Integrity and Fingerprint was missing.
Validation for the same has been added now.
2. In TURNS module When the initial ALLOCATE request is received, the TURNS
stack does not have the necessary information to validate the message using
long term authentication. Hence the request is passed from the TURNS module
to the server application for the validation. The server just checks the db,
makes sure the username exists, and certain basic criteria are met. Further,
the server also generates the HMAC key using the password and passes it to
TURNS stack for authentication of messages. The message integrity check
for the initial ALLOCATE request was missing, the same has been added.

+++ turns/src/turns_table.c
@@ -298,7 +298,7 @@ int32_t turns_table_find_node_for_relayed_transport_address(
     {
         context = &node->context;
         if ((context->protocol == protocol) && 
-                (context->relay_sock == transport_param))
+                (context->relay_sock == (int)transport_param))
         {
             ICE_LOG (LOG_SEV_INFO, "[TURNS] Allocation "\
                     "context found for relayed transport address");
